fix: move createFieldWithPresets.tsx to src/testing
